Beginning update to SWORD 1.8.1

+++ build.sh
@@ -19,17 +19,17 @@ if [ $(uname | grep Darwin) ]; then
 else
   echo Running on Linux
   sudo apt-get update
-  sudo apt-get install -y libtool
-  sudo apt-get install -y autoconf
-  sudo apt-get install -y make
-  sudo apt-get install -y pkg-config
-  sudo apt-get install -y build-essential
-  sudo apt-get install -y subversion
-  sudo apt-get install -y git
-  sudo apt-get install -y zip
-  sudo apt-get install -y firefox
-  # The following is needed on 16.04 Xenial
-  sudo apt-get install -y libtool-bin
+  sudo apt-get install -y build-essential git subversion libtool-bin autoconf make pkg-config zip
+  # install packages for sword
+  sudo apt-get install -y zlib1g-dev libclucene-dev
+  if [[ "$(uname -p)" == "i686" ]]; then
+    # the 32-bit libclucene is installed in an unexpected place for sword-svn, so fix it
+    sudo ln -s /usr/lib/i386-linux-gnu/libclucene-core.so /usr/lib/libclucene-core.so
+  fi
+  # fix bug in clucene
+  sudo sed -i -e 's/#include "CLucene\/clucene-config.h"/#include "CLucene\/CLConfig.h"/' /usr/include/CLucene/SharedHeader.h
+    
+  #sudo apt-get install -y firefox
 fi
 
 # If this is Vagrant, then copy xulsword code locally so as not to 
@@ -67,18 +67,15 @@ find . -name "*.sh" -exec chmod ugo+x {} \;
 
 # COMPILE SWORD ENGINE DEPENDENCIES
 # CLucene
-if [ ! -e "$XULSWORD/Cpp/clucene-core-0.9.21b" ]; then
+if [ $(uname | grep Darwin) ] && [ ! -e "$XULSWORD/Cpp/clucene-core-0.9.21b" ]; then
+  # Patch a problem with MAC compile of clucene
   echo Compiling $XULSWORD/Cpp/clucene-core-0.9.21b
   cd "$XULSWORD/Cpp"
   wget http://sourceforge.net/projects/clucene/files/clucene-core-stable/0.9.21b/clucene-core-0.9.21b.tar.bz2/download
   tar -xf download 
   rm download
-
   cd "$XULSWORD/Cpp/clucene-core-0.9.21b"
-  # Patch a problem with MAC compile of clucene
-  if [ $(uname | grep Darwin) ]; then
-    cp -f ../cluceneMK/osx/repl_tchar.h ./src/CLucene/config
-  fi
+  cp -f ../cluceneMK/osx/repl_tchar.h ./src/CLucene/config
   make clean
   ./configure --disable-multithreading
   make
@@ -86,54 +83,25 @@ if [ ! -e "$XULSWORD/Cpp/clucene-core-0.9.21b" ]; then
   sudo ldconfig
 fi
 
-: <<'COMMENT'
-cd $XULSWORD/Cpp
-# Compile XZ Utils
-if [ ! -e ./xz-5.0.5 ]; then
-	wget http://tukaani.org/xz/xz-5.0.5.tar.bz2
-	tar -xf xz-5.0.5.tar.bz2 
-	rm xz-5.0.5.tar.bz2
-fi
-cd xz-5.0.5
-make clean
-./configure
-make
-sudo make install
-sudo ldconfig
-
-# Compile bzip2
-cd $XULSWORD/Cpp
-if [ ! -e ./clucene-core-0.9.21b ]; then
-	wget http://www.bzip.org/1.0.6/bzip2-1.0.6.tar.gz
-	tar -xf bzip2-1.0.6.tar.gz 
-	rm bzip2-1.0.6.tar.gz
-fi
-cd bzip2-1.0.6
-make clean
-make
-sudo make install
-sudo ldconfig
-COMMENT
-
 # Compile the SWORD engine (at specific rev)
-swordRev=3203
+# 3563 is sword-1.8.1
+swordRev=3563
 if [ ! -e "$XULSWORD/Cpp/sword-svn" ]; then
   cd "$XULSWORD/Cpp"
   svn checkout -r $swordRev http://crosswire.org/svn/sword/trunk sword-svn
   cd sword-svn
-  # fix a SWORD bug or two
-  perl -p -i -e 's/(PKG_CHECK_MODULES\(\[CLUCENE2\], \[libclucene\-core >= 2.3\],,true\))/\#$1/' ./configure.ac
-  cp "$XULSWORD/Cpp/swordMK/osisheadings.cpp" "$XULSWORD/Cpp/sword-svn/src/modules/filters/"
   if [ $(uname | grep Darwin) ]; then
     # use brew's glibtoolize instead of libtoolize
     perl -p -i -e 's/^(LTIZE="\$AUTODIR"")(libtoolize")/$1g$2/' ./autogen.sh
   fi
   ./autogen.sh
-  ./configure --without-icu --without-xz --without-bzip2
+  ./configure
   make
   sudo make install
   sudo ldconfig
 fi
+
+exit
 
 # Download xulrunner (unless it exists already)
 xulrunnerRev=41.0b9
@@ -142,7 +110,7 @@ if [ ! -e "$XULSWORD/xulrunner" ]; then
   if [ $(uname | grep Darwin) ]; then
     xulrunner=xulrunner-$xulrunnerRev.en-US.mac.tar.bz2
   else
-    if [[ "$(uname -m)" == *"i686"* ]]; then
+    if [[ "$(uname -p)" == "i686" ]]; then
       xulrunner=xulrunner-$xulrunnerRev.en-US.linux-i686.tar.bz2
     else
       xulrunner=xulrunner-$xulrunnerRev.en-US.linux-x86_64.tar.bz2
